move : create exception.domain package
