+++ src/main/java/kuit/project/beering/service/FavoriteService.java
@@ -8,8 +8,8 @@ import kuit.project.beering.repository.FavoriteRepository;
 import kuit.project.beering.repository.MemberRepository;
 import kuit.project.beering.repository.drink.DrinkRepository;
 import kuit.project.beering.util.BaseResponseStatus;
-import kuit.project.beering.util.exception.DrinkException;
-import kuit.project.beering.util.exception.MemberException;
+import kuit.project.beering.util.exception.domain.DrinkException;
+import kuit.project.beering.util.exception.domain.MemberException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
