+++ src/main/java/kuit/project/beering/service/ReviewOptionService.java
@@ -5,7 +5,7 @@ import kuit.project.beering.domain.ReviewOption;
 import kuit.project.beering.dto.response.reviewOption.ReviewOptionReadResponseDto;
 import kuit.project.beering.repository.ReviewOptionRepository;
 import kuit.project.beering.repository.drink.DrinkRepository;
-import kuit.project.beering.util.exception.DrinkException;
+import kuit.project.beering.util.exception.domain.DrinkException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
