+++ src/main/java/kuit/project/beering/service/ReviewService.java
@@ -11,9 +11,9 @@ import kuit.project.beering.dto.response.review.ReviewReadResponseDto;
 import kuit.project.beering.dto.response.review.ReviewResponseDto;
 import kuit.project.beering.repository.*;
 import kuit.project.beering.repository.drink.DrinkRepository;
-import kuit.project.beering.util.exception.DrinkException;
-import kuit.project.beering.util.exception.MemberException;
-import kuit.project.beering.util.exception.ReviewException;
+import kuit.project.beering.util.exception.domain.DrinkException;
+import kuit.project.beering.util.exception.domain.MemberException;
+import kuit.project.beering.util.exception.domain.ReviewException;
 import kuit.project.beering.util.s3.AwsS3Uploader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
