+++ src/main/java/kuit/project/beering/service/TabomService.java
@@ -11,8 +11,8 @@ import kuit.project.beering.repository.MemberRepository;
 import kuit.project.beering.repository.ReviewRepository;
 import kuit.project.beering.repository.TabomRepository;
 import kuit.project.beering.util.BaseResponseStatus;
-import kuit.project.beering.util.exception.MemberException;
-import kuit.project.beering.util.exception.ReviewException;
+import kuit.project.beering.util.exception.domain.MemberException;
+import kuit.project.beering.util.exception.domain.ReviewException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
