+++ src/main/java/kuit/project/beering/util/exception/domain/MemberException.java
@@ -1,4 +1,4 @@
-package kuit.project.beering.util.exception;
+package kuit.project.beering.util.exception.domain;
 
 import kuit.project.beering.util.BaseResponseStatus;
 import lombok.Getter;
@@ -6,9 +6,9 @@ import lombok.extern.slf4j.Slf4j;
 
 @Getter
 @Slf4j
-public class FavoriteException extends DomainException {
+public class MemberException extends DomainException {
 
-    public FavoriteException(BaseResponseStatus status) {
+    public MemberException(BaseResponseStatus status) {
         super(status);
         log.info("{} - message : {}", this.getClass().getSimpleName(), status.getResponseMessage());
     }
