+++ src/main/java/kuit/project/beering/util/exception/domain/TabomException.java
@@ -1,4 +1,4 @@
-package kuit.project.beering.util.exception;
+package kuit.project.beering.util.exception.domain;
 
 import kuit.project.beering.util.BaseResponseStatus;
 import lombok.Getter;
@@ -6,9 +6,9 @@ import lombok.extern.slf4j.Slf4j;
 
 @Getter
 @Slf4j
-public class ReviewException extends DomainException {
+public class TabomException extends DomainException {
 
-    public ReviewException(BaseResponseStatus status) {
+    public TabomException(BaseResponseStatus status) {
         super(status);
         log.info("{} - message : {}", this.getClass().getSimpleName(), status.getResponseMessage());
     }
